[Fix] LTX23ConditionPipeline: cast condition tensors to VAE dtype before encode
Same VAE/conv dtype regression as the I2V image path (72c56e88): the
condition_tensor coming out of _preprocess_conditions is in the latent
dtype (typically float32), but self.vae expects bf16, so the engine
warmup crashes with "Input type (float) and bias type (c10::BFloat16)
should be the same" as soon as any condition is supplied.

Cast each condition_tensor to self.vae.dtype before vae.encode; the
condition_latent is already re-cast to the latent dtype right after.

Cover-up tests in test_ltx2_3_pipeline.py:
  - test_i2v_prepare_image_to_video_latents_casts_image_to_vae_dtype:
    asserts the I2V image path now casts to vae.dtype (regression guard
    for the prior fix).
  - test_prepare_condition_latents_casts_condition_to_vae_dtype: same
    guard for the Condition path.
  - In test_forward_cfg_parallel_steps_video_and_audio_scheduler, assert
    that the base forward passes a scalar audio_timestep to the
    transformer, so the per-token video timestep no longer leaks into the
    audio adaLN path (the second regression this branch carried).

+++ vllm_omni/diffusion/models/ltx2/pipeline_ltx2_3.py
@@ -2150,8 +2150,11 @@ class LTX23ConditionPipeline(LTX23Pipeline):
             )
             condition_latents: list[torch.Tensor] = []
             for condition_tensor in condition_frames:
+                # Cast to VAE dtype here; condition_latent is re-cast to `dtype` after encode.
                 condition_latent = retrieve_latents(
-                    self.vae.encode(condition_tensor), generator=generator, sample_mode="argmax"
+                    self.vae.encode(condition_tensor.to(dtype=self.vae.dtype)),
+                    generator=generator,
+                    sample_mode="argmax",
                 )
                 condition_latent = self._normalize_latents(
                     condition_latent, self.vae.latents_mean, self.vae.latents_std
